// src/app/openapi/model/catalog-item-filter.ts
/**
 * Component Catalog REST API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CatalogItemFilter { 
    label?: string;
    options?: Set<string>;
}